WIP support 3:1 aspect ratio for event cards for all device sizes
